Moved new features from "io/input.py" to "helpers/python.py" to be consistent with the mirror functions for s3 bucket
